chore: renaming functions and variables
