+++ src/pages/ListPage.tsx
@@ -10,7 +10,7 @@ import mikuBgUrl from '../assets/miku/miku.jpg';
 import EateryCardGrid from '../components/EateryCardGrid';
 import Drawer from '../components/Drawer';
 import { DrawerAPIContextProvider, useDrawerAPIContext } from '../contexts/DrawerAPIContext';
-import { type SortOption, useFilteredLocations, useSortedLocations } from '../util/useFilteredLocations';
+import { type SortOption, useFilteredLocations, useSortedLocations } from '../util/useLocationList';
 import './ListPage.css';
 import { CardViewPreference } from '../util/storage';
 import Footer from '../components/Footer';
